requirements.txt
requests==2.28.2
beautifulsoup4==4.12.2
python-telegram-bot==20.0
apscheduler==3.10.1
ftplib
python-dotenv==1.0.0